Unify error codes and rename RPC wrapper
