+++ pkgs/standards/peagen/peagen/gateway/rpc/secrets.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from .. import Session, dispatcher, log
 from peagen.defaults import SECRETS_ADD, SECRETS_GET, SECRETS_DELETE
 from ..db_helpers import delete_secret, fetch_secret, upsert_secret
-from peagen.defaults.error_codes import ErrorCode
+from peagen.protocols.error_codes import Code as ErrorCode
 from peagen.transport.jsonrpc import RPCException
