+++ pkgs/standards/peagen/peagen/gateway/rpc/tasks.py
@@ -4,7 +4,7 @@ import json
 import uuid
 import typing as t
 from peagen.transport.jsonrpc import RPCException
-from peagen.defaults.error_codes import ErrorCode
+from peagen.protocols.error_codes import Code as ErrorCode
 from peagen.defaults import (
     TASK_SUBMIT,
     TASK_CANCEL,
@@ -52,6 +52,7 @@ def _parse_task_create(task: t.Any) -> TaskCreate:
     if not isinstance(task, TaskCreate):
         raise TypeError("TaskCreate required")
     return task
+
 
 # --------------Basic Task Methods ---------------------------------
