+++ pkgs/standards/peagen/peagen/transport/__init__.py
@@ -4,12 +4,12 @@ from peagen.protocols import (
     Response as RPCResponse,
     Error as RPCErrorData,
 )
-from peagen.transport.schemas import RPCError
+from peagen.transport.schemas import RPCException
 
 __all__ = [
     "RPCDispatcher",
     "RPCRequest",
     "RPCResponse",
-    "RPCError",
+    "RPCException",
     "RPCErrorData",
 ]
